+++ src/main/java/eth/epieffe/jwalker/Move.java
@@ -6,11 +6,11 @@ public class Move<T> {
 
     public final String move;
 
-    public final int cost;
+    public final double cost;
 
     public final T config;
 
-    public Move(String move, int cost, T config) {
+    public Move(String move, double cost, T config) {
         this.move = move;
         this.cost = cost;
         this.config = config;
